move storage package and relevant code for managing files on storage to generic indexshipper
