+++ pkg/storage/stores/indexshipper/storage/cached_client.go
