+++ pkg/storage/stores/indexshipper/storage/client.go
