+++ pkg/storage/stores/indexshipper/storage/prefixed_object_client.go
