+++ pkg/storage/stores/indexshipper/uploads/table.go
@@ -8,7 +8,7 @@ import (
 	"github.com/go-kit/log"
 
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper/index"
-	"github.com/grafana/loki/pkg/storage/stores/shipper/storage"
+	"github.com/grafana/loki/pkg/storage/stores/indexshipper/storage"
 	util_log "github.com/grafana/loki/pkg/util/log"
 )
