+++ pkg/storage/stores/indexshipper/uploads/table_manager_test.go
@@ -11,7 +11,7 @@ import (
 
 	"github.com/grafana/loki/pkg/storage/chunk/client/local"
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper/index"
-	"github.com/grafana/loki/pkg/storage/stores/shipper/storage"
+	"github.com/grafana/loki/pkg/storage/stores/indexshipper/storage"
 )
 
 const objectsStorageDirName = "objects"
