+++ pkg/storage/stores/shipper/index/compactor/table_compactor_test.go
@@ -22,9 +22,8 @@ import (
 	"github.com/grafana/loki/pkg/storage/chunk/client/util"
 	"github.com/grafana/loki/pkg/storage/config"
 	"github.com/grafana/loki/pkg/storage/stores/indexshipper/compactor"
-	"github.com/grafana/loki/pkg/storage/stores/shipper/storage"
+	"github.com/grafana/loki/pkg/storage/stores/indexshipper/storage"
 	"github.com/grafana/loki/pkg/storage/stores/shipper/testutil"
-	shipper_util "github.com/grafana/loki/pkg/storage/stores/shipper/util"
 	util_log "github.com/grafana/loki/pkg/util/log"
 )
 
@@ -85,14 +84,14 @@ func (m *mockIndexSet) ListSourceFiles() []storage.IndexFile {
 }
 
 func (m *mockIndexSet) GetSourceFile(indexFile storage.IndexFile) (string, error) {
-	decompress := shipper_util.IsCompressedFile(indexFile.Name)
+	decompress := storage.IsCompressedFile(indexFile.Name)
 	dst := filepath.Join(m.workingDir, indexFile.Name)
 	if decompress {
 		dst = strings.Trim(dst, ".gz")
 	}
 
-	err := shipper_util.DownloadFileFromStorage(dst, shipper_util.IsCompressedFile(indexFile.Name),
-		false, shipper_util.LoggerWithFilename(util_log.Logger, indexFile.Name),
+	err := storage.DownloadFileFromStorage(dst, storage.IsCompressedFile(indexFile.Name),
+		false, storage.LoggerWithFilename(util_log.Logger, indexFile.Name),
 		func() (io.ReadCloser, error) {
 			rc, _, err := m.objectClient.GetObject(context.Background(), path.Join(m.tableName, m.userID, indexFile.Name))
 			return rc, err
